feat: Add SiteType CRUD operations
- Implemented SiteTypeController with REST endpoints for CRUD operations
- Added methods to handle create, read, update, and delete operations
- Included proper response handling with DataApiResponse
- Added input validation for create and update operations
- Implemented proper error handling and logging

+++ src/main/java/com/sievex/crawler/repository/JobTypeRepository.java
@@ -3,6 +3,6 @@ package com.sievex.crawler.repository;
 import com.sievex.crawler.entity.JobType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobTypeRepository extends JpaRepository<JobType, Long> {
+public interface JobTypeRepository extends JpaRepository<JobType, Byte> {
     JobType findByAlias(String alias);
 }
